pyproject.toml
[build-system]
  build-backend = "uv_build"
  requires = ["uv_build"]

[project]
  authors = [{ email = "d.wan@icloud.com", name = "Derek Wan" }]
  dependencies = [
    "click >=8.3.1, <8.4",
    "dycw-utilities >=0.170.0, <0.171",
    "rich>=14.2.0, <14.3",
    "typed-settings[attrs, click] >=25.3.0, <25.4",
  ]
  description = "Tag the latest commit with the version read from `bump-my-version`"
  name = "dycw-tag-commit"
  readme = "README.md"
  requires-python = ">= 3.13"
  version = "0.5.0"

  [project.scripts]
    tag-commit = "tag_commit.cli:_main"

[dependency-groups]
  dev = ["dycw-utilities[test]", "pyright", "rich"]

[tool]
  [tool.uv]
    [tool.uv.build-backend]
      module-name = "tag_commit"
      module-root = "src"
